Split SMTSolverImpl wrappers into cpp file
